requirements.txt
flask>=2.3.3
python-dotenv>=1.0.0
psycopg2-binary>=2.9.7
flask-sqlalchemy>=3.1.1
flask-wtf>=1.1.1
flask-session>=0.5.0
flask-bcrypt>=1.0.1
flask-mail>=0.9.1
flask-cors>=4.0.0
authlib>=1.2.1
flask-jwt-extended>=4.5.3
requests>=2.31.0
gunicorn>=21.2.0